+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/LabeledMetricType.kt
@@ -7,6 +7,7 @@ package mozilla.telemetry.glean.private
 import androidx.annotation.VisibleForTesting
 import mozilla.telemetry.glean.internal.LabeledBoolean
 import mozilla.telemetry.glean.internal.LabeledCounter
+import mozilla.telemetry.glean.internal.LabeledQuantity
 import mozilla.telemetry.glean.internal.LabeledString
 import mozilla.telemetry.glean.testing.ErrorType
 
@@ -50,6 +51,7 @@ class LabeledMetricType<T>(
             is CounterMetricType -> LabeledCounter(meta, labels?.toList())
             is BooleanMetricType -> LabeledBoolean(meta, labels?.toList())
             is StringMetricType -> LabeledString(meta, labels?.toList())
+            is QuantityMetricType -> LabeledQuantity(meta, labels?.toList())
             else -> error("Can not create a labeled version of this metric type")
         }
     }
@@ -77,6 +79,7 @@ class LabeledMetricType<T>(
             is LabeledCounter -> this.inner.get(label) as T
             is LabeledBoolean -> this.inner.get(label) as T
             is LabeledString -> this.inner.get(label) as T
+            is LabeledQuantity -> this.inner.get(label) as T
             else -> error("Can not create a labeled version of this metric type")
         }
     }
@@ -114,6 +117,7 @@ class LabeledMetricType<T>(
             is LabeledCounter -> this.inner.testGetNumRecordedErrors(errorType)
             is LabeledBoolean -> this.inner.testGetNumRecordedErrors(errorType)
             is LabeledString -> this.inner.testGetNumRecordedErrors(errorType)
+            is LabeledQuantity -> this.inner.testGetNumRecordedErrors(errorType)
             else -> error("Can not create a labeled version of this metric type")
         }
     }