+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 * Android
    * Updated to Android NDK r28b ([#3118](https://github.com/mozilla/glean/pull/3118))
+* Kotlin
+   * Added support for labeled quantity metric type ([#3121](https://github.com/mozilla/glean/pull/3121))
+* Swift
+   * Added support for labeled quantity metric type ([#3121](https://github.com/mozilla/glean/pull/3121))
+* Python
+   * Added support for labeled quantity metric type ([#3121](https://github.com/mozilla/glean/pull/3121))
 
 [Full changelog](https://github.com/mozilla/glean/compare/v64.2.0...main)